models: split ParamResolver from ParamStore
- Keep ParamStore runtime-light (storage-only).

- Add I18nResolver + ComputedValueEvaluator for asset-driven labels and computed STATUS.

- Update CLI to use ParamResolver and add --lang (ENV: PYBO_LANG).

- Adjust tests to target resolver/evaluator instead of ParamStore heavy helpers.
