+++ src/pybragerone/models/__init__.py
@@ -6,10 +6,12 @@ from .events import EventBus
 from .menu import MenuMeta, MenuParameter, MenuParameters, MenuResult, MenuRoute
 from .menu_manager import MenuManager, MenuProcessor, RawMenuData
 from .param import ParamFamilyModel, ParamStore
+from .param_resolver import ComputedValueEvaluator, ParamResolver, ResolvedValue
 from .token import CLITokenStore, Token, TokenStore
 
 __all__ = [
     "CLITokenStore",
+    "ComputedValueEvaluator",
     "EventBus",
     "LiveAssetsCatalog",
     "MenuManager",
@@ -20,8 +22,10 @@ __all__ = [
     "MenuResult",
     "MenuRoute",
     "ParamFamilyModel",
+    "ParamResolver",
     "ParamStore",
     "RawMenuData",
+    "ResolvedValue",
     "Token",
     "TokenStore",
     "api",
